moved detectPose useEffect and webcamCanvas to ExerciseBox component, generalized SquatPage but not others
