+++ app/src/components/ExerciseBox.js
@@ -1,7 +1,9 @@
-import React, { useState, useRef } from "react";
-import { Button, Box, Typography } from "@mui/material";
+import React, { useState, useEffect, useRef } from "react";
+import { Box, Typography } from "@mui/material";
+import WebcamCanvas from "./WebcamCanvas";
 import VideoCanvas from "./VideoCanvas";
 import startPoseDetection from "../utils/PoseDetectorPoseVideo";
+import detectPose from "../utils/PoseDetector";
 
 /**
  * A reusable layout component for exercise tracking pages.
@@ -13,10 +15,23 @@ import startPoseDetection from "../utils/PoseDetectorPoseVideo";
  *
  * @returns {JSX.Element} The JSX code for the ExerciseBox layout.
  */
-function ExerciseBox({ title, webcamCanvas, feedbackPanel, processPoseResults }) {
+function ExerciseBox({ title, feedbackPanel, processPoseResults, targetAngles }) {
+    const webcamRef = useRef(null);
+    const canvasRef = useRef(null);
+    const [dimensions] = useState({
+        width: window.innerWidth,
+        height: window.innerHeight,
+    });
+
     const videoRef = useRef(null);
     const videoCanvasRef = useRef(null);
     const [useVideo, setUseVideo] = useState(false);
+
+    useEffect(() => {
+        detectPose(webcamRef, canvasRef, processPoseResults);
+
+        return () => { };
+    }, [targetAngles]);
 
     const handleVideoUpload = (event) => {
         const file = event.target.files[0];
@@ -58,7 +73,10 @@ function ExerciseBox({ title, webcamCanvas, feedbackPanel, processPoseResults })
                 }}
             >
                 <Box sx={{ display: useVideo ? "none" : "" }}>
-                    {webcamCanvas}
+                    <WebcamCanvas
+                        dimensions={dimensions}
+                        ref={{ webcamRef: webcamRef, canvasRef: canvasRef }}
+                    />
                 </Box>
                 <Box sx={{ display: useVideo ? "" : "none" }}>
                     <VideoCanvas
